pyproject.toml
[tool.poetry]
name = "groq-api"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.112.0"
groq = "^0.9.0"
asyncio = "^3.4.3"
requests = "^2.32.3"
ruff = "^0.5.7"
uvicorn = "^0.30.5"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
